Add compatibility with platform that don't have getrusage(2)
For platform without getrusage support, we rely on postgres
implementation which only supports user and system time.  The other
fields will still be present, to keep a consistent definition across all
platforms, but will be NULL.

While at it, also gather all counters in a single struct, and keep the
mutex only to copy this struct in local varibale in pg_stat_kcache() SRF
to increase concurrency.